# indexer/openapi.yaml
openapi: 3.0.0
info:
  title: ElcareHub Indexer API
  description: Off-chain event indexer and REST API for the ElcareHub NFT
    marketplace on Stellar Soroban. All BigInt values (IDs, endTime) are
    serialised as decimal strings.
  version: 1.0.0
  contact:
    name: ElcareHub
    url: https://elcarehub.io
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://indexer.elcarehub.io
    description: Production
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Listing not found
          required:
            - code
            - message
      required:
        - error
  # ── Conditional GET headers (Issue #508) ──────────────────────────────────
  # Reusable header and response definitions for ETag / conditional GET support.
  # Applied to all cacheable endpoints so clients can implement HTTP caching
  # without guessing at the correct header names or value formats.
  headers:
...
    Vary:
      description: >
        Instructs CDN / proxy caches to key on Accept-Encoding in addition to
        the URL, so compressed and uncompressed variants are stored separately.
      schema:
        type: string
        example: Accept-Encoding
...
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: >
        Client's cached ETag value (obtained from a previous response's `ETag`
        header).  When the server's current ETag matches, it returns **304 Not
        Modified** with no body instead of re-executing the query.
      schema:
        type: string
        example: '"a3f8c2d1e4b7a9f0c2d4e6b8a0c1d3e5"'
paths:
  /listings:
    get:
      tags:
        - Listings
      summary: List all listings
      description: Returns listings with optional filters. When `limit` or `offset`
        are supplied the response wraps results in a pagination envelope.
      parameters:
        - schema:
            type: string
            description: Filter by artist Stellar address
          required: false
          description: Filter by artist Stellar address
          name: artist
          in: query
        - schema:
            type: string
            description: Filter by current owner address
          required: false
          description: Filter by current owner address
          name: owner
          in: query
        - schema:
            type: string
            enum:
              - Active
              - Sold
              - Cancelled
              - Auction
            description: Filter by listing status
          required: false
          description: Filter by listing status
          name: status
          in: query
        - schema:
            type: string
            description: Full-text search on artist address or collection
          required: false
          description: Full-text search on artist address or collection
          name: search
          in: query
        - schema:
            type: number
            nullable: true
            minimum: 0
            description: Minimum price (inclusive)
          required: false
          description: Minimum price (inclusive)
          name: minPrice
          in: query
        - schema:
            type: number
            nullable: true
            minimum: 0
            description: Maximum price (inclusive)
          required: false
          description: Maximum price (inclusive)
          name: maxPrice
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            maximum: 1000
            description: Max results to return (max 1000)
          required: false
          description: Max results to return (max 1000)
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            maximum: 10000
            description: Number of results to skip
          required: false
          description: Number of results to skip
          name: offset
          in: query
      responses:
        "200":
          description: Listing array, or paginated envelope when limit/offset are used
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                  - type: object
                    properties:
                      listings:
                        type: array
                        items:
                          $ref: "#/components/schemas/Listing"
                      total:
                        type: integer
                    required:
                      - listings
                      - total
  /listings/{id}:
    get:
      tags:
        - Listings
      summary: Get a single listing
      parameters:
        - schema:
            type: string
            description: Listing ID
          required: true
          description: Listing ID
          name: id
          in: path
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /listings/{id}/history:
    get:
      tags:
        - Listings
      summary: Get on-chain event history for a listing
      parameters:
        - schema:
            type: string
            description: Listing ID
          required: true
          description: Listing ID
          name: id
          in: path
      responses:
        "200":
          description: Ordered list of marketplace events for this listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceEvent"
        "400":
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auctions:
    get:
      tags:
        - Auctions
      summary: List all auctions
      parameters:
        - schema:
            type: string
            description: Filter by creator address
          required: false
          description: Filter by creator address
          name: creator
          in: query
        - schema:
            type: string
            enum:
              - Active
              - Finalized
              - Cancelled
            description: Filter by auction status
          required: false
          description: Filter by auction status
          name: status
          in: query
      responses:
        "200":
          description: Auction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Auction"
  /auctions/{id}:
    get:
      tags:
        - Auctions
      summary: Get a single auction
      parameters:
        - schema:
            type: string
            description: Auction ID
          required: true
          description: Auction ID
          name: id
          in: path
      responses:
        '200':
          description: Auction details with bid history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Auction'
                  - type: object
                    properties:
                      bids:
                        type: array
                        description: Bid history, newest first
                        items:
                          $ref: '#/components/schemas/Bid'
        '404':
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offers:
    get:
      tags:
        - Offers
      summary: List offers
      description: Returns all offers. Use `listing_id` to filter to a specific listing.
      parameters:
        - schema:
            type: string
            pattern: ^\d+$
            description: Filter by listing ID (numeric string)
          required: false
          description: Filter by listing ID (numeric string)
          name: listing_id
          in: query
      responses:
        "200":
          description: Offer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /activity/recent:
    get:
      tags:
        - Activity
      summary: Get the 20 most recent marketplace events
      description: >
        Returns the latest cross-marketplace activity feed. Cached for 30 s.

        **Conditional GET** — Supports `If-None-Match` / ETag revalidation.
        Because this endpoint includes provisional (unconfirmed) events that
        can be rolled back by a chain reorg, the `Cache-Control` header is
        `no-cache`: clients must revalidate on every request, but a valid
        `If-None-Match` match still returns **304** with no body and avoids
        the DB query.
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        "200":
          description: Recent events
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Vary:
              $ref: '#/components/headers/Vary'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceEvent"
        "304":
          $ref: '#/components/responses/NotModified'
  /collections:
    get:
      tags:
        - Collections
      summary: List all deployed collections
      description: Cached for 60 s.
      parameters:
        - schema:
            type: string
            description: Filter by collection type (normal_721, normal_1155, lazy_721,
              lazy_1155)
          required: false
          description: Filter by collection type (normal_721, normal_1155, lazy_721,
            lazy_1155)
          name: kind
          in: query
        - schema:
            type: string
            description: Filter by creator address
          required: false
          description: Filter by creator address
          name: creator
          in: query
      responses:
        "200":
          description: Collection list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
  /creators/{address}/collections:
    get:
      tags:
        - Collections
      summary: Get all collections deployed by a creator
      parameters:
        - schema:
            type: string
            description: Creator Stellar address
          required: true
          description: Creator Stellar address
          name: address
          in: path
      responses:
        "200":
          description: Collections by creator
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
  /collections/{address}:
    get:
      tags:
        - Collections
      summary: Get collection detail with marketplace stats and recent activity
      description: Returns the collection metadata plus aggregated marketplace statistics (active listings, sales volume, floor price, unique owners/tokens) and the latest events involving this collection's listings. Cached for 60 s.
      parameters:
        - schema:
            type: string
            description: Collection contract address
          required: true
          description: Collection contract address
          name: address
          in: path
      responses:
        "200":
          description: Collection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetail"
        "404":
          $ref: '#/components/responses/NotFound'
...
  /wallets/{address}/activity:
    get:
      tags:
        - Wallets
      summary: Get activity feed for a wallet
      description: >
        Returns events where the address is the `actor` or appears in the
        event JSON payload (buyer, artist, offerer, bidder, winner, creator).
        Rate-limited to 20 req/min.

        **Conditional GET** — Supports `If-None-Match` / ETag revalidation.
        `Cache-Control: no-cache` is set because the feed contains provisional
        events subject to reorg rollback; clients must revalidate but a
        matching `If-None-Match` avoids the DB query and returns **304**.
      parameters:
        - schema:
            type: string
            description: Wallet Stellar address
          required: true
          description: Wallet Stellar address
          name: address
          in: path
        - schema:
            type: integer
            nullable: true
            minimum: 0
            maximum: 200
            description: Max results (default 50, max 200)
          required: false
          description: Max results (default 50, max 200)
          name: limit
          in: query
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        "200":
          description: Wallet event feed
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Vary:
              $ref: '#/components/headers/Vary'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceEvent"
        "304":
          $ref: '#/components/responses/NotModified'
  /wallets/{address}/royalty-stats:
    get:
      tags:
        - Wallets
      summary: Get royalty earnings summary for an artist
      description: Calculates total royalties earned from secondary sales.
        Rate-limited to 20 req/min.
      parameters:
        - schema:
            type: string
            description: Artist Stellar address
          required: true
          description: Artist Stellar address
          name: address
          in: path
      responses:
        "200":
          description: Royalty statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoyaltyStats"
...
  /events:
    get:
      tags:
        - System
      summary: Server-Sent Events stream
      description: Real-time event stream using SSE. Supports `Last-Event-Id` header
        for replay of up to the last 200 events. Returns 503 when the connection
        limit is reached.
      responses:
        "200":
          description: SSE stream (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: string
        "503":
          description: Too many SSE connections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sync/gaps/{id}:
    get:
      summary: Get a single ledger gap
      tags: [Sync]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Gap details including repair job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerGap'
        '404':
          description: Gap not found
...